Extract ROS object detection logic out of front and down cam nodes and into separate file
