+++ ros2_ws/src/vision/launch/object_detection.launch.xml
@@ -2,14 +2,14 @@
     <arg name="front_model" default="/home/douglas/AUV-2026/ros2_ws/src/vision/model_pipeline/models/frontcam.pt"/>
     <arg name="down_model" default="/home/douglas/AUV-2026/ros2_ws/src/vision/model_pipeline/models/downcam.pt"/>
 
-    <node pkg="vision" exec="front_cam_object_detection.py" name="front_cam_object_detection">
+    <node pkg="vision" exec="front_cam_object_detection" name="front_cam_object_detection">
         <param name="class_names" value="['gate', 'lane_marker', 'red_pipe', 'white_pipe', 'octagon', 'table', 'bin', 'board', 'shark', 'sawfish']"/>
         <param name="model_path" value="$(var front_model)"/>
         <param name="input_topic" value="/zed/zed_node/rgb/image_rect_color"/>
         <param name="output_topic" value="/vision/front_cam/detections"/>
         <param name="queue_size" value="10"/>
     </node>
-    <node pkg="vision" exec="down_cam_object_detection.py" name="down_cam_object_detection">
+    <node pkg="vision" exec="down_cam_object_detection" name="down_cam_object_detection">
         <param name="class_names" value="['gate', 'octagon_table', 'octagon_top', 'path_marker', 'sawfish', 'shark']"/>
         <param name="model_path" value="$(var down_model)"/>
         <param name="input_topic" value="/zed/zed_node/rgb/image_rect_color"/>
