+++ ros2_ws/src/vision/package.xml
@@ -12,6 +12,7 @@
 
   <depend>rclpy</depend>
   <depend>sensor_msgs</depend>
+  <depend>vision_msgs</depend>
   <depend>cv_bridge</depend>
 
   <test_depend>ament_copyright</test_depend>
